Binance integration
Former-commit-id: 071dddaefa2f09fa47bb8fdbc49b865ce66ee178

+++ HFTPlatform/zHFT.InstructionBasedMarketClient.BitMex/BitMexInstructionBasedMarketClient.cs
@@ -326,7 +326,7 @@ namespace zHFT.InstructionBasedMarketClient.BitMex
                     WSMarketDataManager.SubscribeEvents(BaseManager._TRADE, UpdateTrade);
 
                     ActiveSecurities = new Dictionary<long, Security>();
-                    ContractsTimeStamps = new Dictionary<int, DateTime>();
+                    ContractsTimeStamps = new Dictionary<long, DateTime>();
                     return true;
                 }
                 else
